backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
groq==0.9.0
chromadb==0.4.24
sentence-transformers==3.0.0
python-dotenv==1.0.1
huggingface-hub==0.23.4